Add the timeline
Adds the timeline page, where the latest events of interest for the community of
users are displayed. More specifically, new and updated kudos, seasons, groups
and prizes are shown as a list. The elements of such list have a class to be
able to easily apply different styles depending on the event type.